style: 도메인 별로 디렉토리 분리
