--- src/main/java/roomescape/ReservationTime.java
@@ -1,16 +1,0 @@
-package roomescape;
-
-import java.time.LocalTime;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NonNull;
-
-@Getter
-@AllArgsConstructor
-public class ReservationTime {
-    @NonNull
-    private Long id;
-
-    @NonNull
-    private LocalTime startAt;
-}
