--- src/main/java/roomescape/controller/ReservationController.java
@@ -1,42 +1,0 @@
-package roomescape.controller;
-
-import jakarta.validation.Valid;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.ReservationRequest;
-import roomescape.service.ReservationService;
-import roomescape.service.dto.ReservationResponse;
-
-@RestController
-@RequestMapping("/reservations")
-@RequiredArgsConstructor
-public class ReservationController {
-
-    private final ReservationService reservationService;
-
-    @GetMapping
-    public List<ReservationResponse> getReservations() {
-        return reservationService.getReservations();
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ReservationResponse createReservation(@Valid @RequestBody ReservationRequest request) {
-        return reservationService.saveReservation(request);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteReservation(@PathVariable Long id) {
-        reservationService.deleteReservation(id);
-    }
-}
