--- src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,42 +1,0 @@
-package roomescape.controller;
-
-import jakarta.validation.Valid;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.CreateTimeRequest;
-import roomescape.service.ReservationTimeService;
-import roomescape.service.dto.ReservationTimeResponse;
-
-@RestController
-@RequestMapping("/times")
-@RequiredArgsConstructor
-public class ReservationTimeController {
-
-    private final ReservationTimeService reservationTimeService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ReservationTimeResponse saveTime(@Valid @RequestBody CreateTimeRequest createTimeRequest) {
-        return reservationTimeService.saveTime(createTimeRequest.startAt());
-    }
-
-    @GetMapping
-    public List<ReservationTimeResponse> findAll() {
-        return reservationTimeService.findAll();
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Long id) {
-        reservationTimeService.delete(id);
-    }
-}
