--- src/main/java/roomescape/controller/ViewController.java
@@ -1,23 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ViewController {
-
-    @GetMapping("/admin")
-    public String getAdminPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String getReservationPage() {
-        return "admin/reservation";
-    }
-
-    @GetMapping("/admin/time")
-    public String getTimePage() {
-        return "admin/time";
-    }
-}
