--- src/main/java/roomescape/controller/dto/CreateTimeRequest.java
@@ -1,7 +1,0 @@
-package roomescape.controller.dto;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalTime;
-
-public record CreateTimeRequest(@NotNull LocalTime startAt) {
-}
