--- src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,8 +1,0 @@
-package roomescape.controller.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
-
-public record ReservationRequest(@NotBlank String name, @NotNull LocalDate date, @NotNull Long timeId) {
-}
