+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,0 +1,24 @@
+package roomescape.reservation;
+
+
+import java.time.LocalDate;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NonNull;
+import roomescape.reservationtime.ReservationTime;
+
+@Getter
+@AllArgsConstructor
+public class Reservation {
+    @NonNull
+    private final Long id;
+
+    @NonNull
+    private final String name;
+
+    @NonNull
+    private final LocalDate date;
+
+    @NonNull
+    private final ReservationTime time;
+}
