+++ src/main/java/roomescape/reservation/controller/dto/ReservationRequest.java
@@ -1,0 +1,8 @@
+package roomescape.reservation.controller.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDate;
+
+public record ReservationRequest(@NotBlank String name, @NotNull LocalDate date, @NotNull Long timeId) {
+}
