+++ src/main/java/roomescape/reservation/service/dto/ReservationResponse.java
@@ -1,0 +1,16 @@
+package roomescape.reservation.service.dto;
+
+import java.time.LocalDate;
+import roomescape.reservation.Reservation;
+import roomescape.reservationtime.service.dto.ReservationTimeResponse;
+
+public record ReservationResponse(Long id, String name, LocalDate date, ReservationTimeResponse time) {
+    public ReservationResponse(Reservation reservation) {
+        this(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                new ReservationTimeResponse(reservation.getTime())
+        );
+    }
+}
