+++ src/main/java/roomescape/reservationtime/ReservationTime.java
@@ -1,0 +1,16 @@
+package roomescape.reservationtime;
+
+import java.time.LocalTime;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NonNull;
+
+@Getter
+@AllArgsConstructor
+public class ReservationTime {
+    @NonNull
+    private Long id;
+
+    @NonNull
+    private LocalTime startAt;
+}
