+++ src/main/java/roomescape/reservationtime/controller/dto/CreateTimeRequest.java
@@ -1,0 +1,7 @@
+package roomescape.reservationtime.controller.dto;
+
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalTime;
+
+public record CreateTimeRequest(@NotNull LocalTime startAt) {
+}
