+++ src/main/java/roomescape/reservationtime/repository/ReservationTimeRepository.java
@@ -1,0 +1,17 @@
+package roomescape.reservationtime.repository;
+
+import java.time.LocalTime;
+import java.util.List;
+import java.util.Optional;
+import roomescape.reservationtime.ReservationTime;
+
+public interface ReservationTimeRepository {
+
+    ReservationTime save(final LocalTime startAt);
+
+    List<ReservationTime> findAll();
+
+    void delete(final Long id);
+
+    Optional<ReservationTime> findById(final Long id);
+}
