+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -1,0 +1,32 @@
+package roomescape.reservationtime.service;
+
+import java.time.LocalTime;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import roomescape.reservationtime.ReservationTime;
+import roomescape.reservationtime.repository.ReservationTimeRepository;
+import roomescape.reservationtime.service.dto.ReservationTimeResponse;
+
+@Service
+@RequiredArgsConstructor
+public class ReservationTimeService {
+
+    private final ReservationTimeRepository reservationTimeRepository;
+
+    public ReservationTimeResponse saveTime(final LocalTime time) {
+        final ReservationTime reservationTime = reservationTimeRepository.save(time);
+        return new ReservationTimeResponse(reservationTime);
+    }
+
+    public List<ReservationTimeResponse> findAll() {
+        final List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
+        return reservationTimes.stream()
+                .map(ReservationTimeResponse::new)
+                .toList();
+    }
+
+    public void delete(final Long id) {
+        reservationTimeRepository.delete(id);
+    }
+}
