+++ src/main/java/roomescape/reservationtime/service/dto/ReservationTimeResponse.java
@@ -1,0 +1,14 @@
+package roomescape.reservationtime.service.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+import roomescape.reservationtime.ReservationTime;
+
+public record ReservationTimeResponse(Long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
+    public ReservationTimeResponse(final ReservationTime reservationTime) {
+        this(
+                reservationTime.getId(),
+                reservationTime.getStartAt()
+        );
+    }
+}
