--- src/main/java/roomescape/service/dto/ReservationResponse.java
@@ -1,15 +1,0 @@
-package roomescape.service.dto;
-
-import java.time.LocalDate;
-import roomescape.Reservation;
-
-public record ReservationResponse(Long id, String name, LocalDate date, ReservationTimeResponse time) {
-    public ReservationResponse(Reservation reservation) {
-        this(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate(),
-                new ReservationTimeResponse(reservation.getTime())
-        );
-    }
-}
