--- src/main/java/roomescape/service/dto/ReservationTimeResponse.java
@@ -1,14 +1,0 @@
-package roomescape.service.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalTime;
-import roomescape.ReservationTime;
-
-public record ReservationTimeResponse(Long id, @JsonFormat(pattern = "HH:mm") LocalTime startAt) {
-    public ReservationTimeResponse(final ReservationTime reservationTime) {
-        this(
-                reservationTime.getId(),
-                reservationTime.getStartAt()
-        );
-    }
-}
