--- src/test/java/roomescape/service/stub/StubReservationTimeRepository.java
@@ -1,41 +1,0 @@
-package roomescape.service.stub;
-
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import roomescape.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
-
-public class StubReservationTimeRepository implements ReservationTimeRepository {
-
-    private final List<ReservationTime> data = new ArrayList<>();
-
-    public StubReservationTimeRepository(final ReservationTime... times) {
-        data.addAll(List.of(times));
-    }
-
-    @Override
-    public ReservationTime save(final LocalTime startAt) {
-        ReservationTime newTime = new ReservationTime(3L, startAt);
-        data.add(newTime);
-        return newTime;
-    }
-
-    @Override
-    public List<ReservationTime> findAll() {
-        return List.copyOf(data);
-    }
-
-    @Override
-    public void delete(final Long id) {
-        data.removeIf(time -> time.getId().equals(id));
-    }
-
-    @Override
-    public Optional<ReservationTime> findById(final Long id) {
-        return data.stream()
-                .filter(time -> time.getId().equals(id))
-                .findFirst();
-    }
-}
